Fix bug where themostat in "auto" would show "off" (#103)
* Fix bug where themostat in "auto" would show "off"

`SystemSwitchPosition` has a valid value of `5`, i.e. "auto mode
(cooling). This value was incorrectly intepreted as "off". It should be
interpreted as "auto".

Because my research indicates that a value of `4` is "auto mode
(heating)", I was able to modify some code to show the correct target
temperatures.

This change fixes issue NorthernMan54#70, which itself is a regression
of NorthernMan54#28.

* Revert accidental change